feat: resolve issue #520 (allowlist JSON array) and #521 (consolidate User entities)
Issue #520: Changed feature_flags allowlist column from comma-separated text to
simple-array (JSON array). Updated entity, DTO, service logic, and migration.

Issue #521: Consolidated two separate User entities (auth/entities/user.entity.ts
and users/entities/user.entity.ts) into a single canonical entity at
auth/entities/user.entity.ts with all fields from both. Added UserStatus enum,
licenseExpiryDate, department, licenseVerifiedAt, verifiedBy,
lastAccessRevocationAt, revocationReason, status, and patientProfile
relationship. Made users/entities/user.entity.ts a re-export bridge.
Updated all imports across the codebase to use the canonical path.
Also added patientProfile OneToOne relationship.
