+++ .gitignore
@@ -112,3 +112,4 @@ backup_*.sql
 skills-lock.json
 
 issue.md
+vrickish.md
