+++ src/OAuth2/oidc-identity.entity.ts
@@ -7,7 +7,7 @@ import {
   PrimaryGeneratedColumn,
   UpdateDateColumn,
 } from 'typeorm';
-import { User } from '../../users/entities/user.entity';
+import { User } from '../../auth/entities/user.entity';
 
 /**
  * Stores the binding between an external OIDC identity (provider + subject)
