+++ src/users/dto/update-user.dto.ts
@@ -1,7 +1,7 @@
 import { PartialType } from '@nestjs/mapped-types';
 import { CreateUserDto } from './create-user.dto';
 import { IsBoolean, IsEnum, IsOptional, IsString } from 'class-validator';
-import { UserStatus } from '../enums/medical-role.enum';
+import { UserStatus } from '../../auth/entities/user.entity';
 
 export class UpdateUserDto extends PartialType(CreateUserDto) {
   @IsEnum(UserStatus)
