+++ src/users/entities/patient.entity.ts
@@ -7,7 +7,7 @@ import {
   OneToOne,
   JoinColumn,
 } from 'typeorm';
-import { User } from './user.entity';
+import { User } from '../../auth/entities/user.entity';
 
 @Entity('patients')
 export class Patient {
